scripts/requirements.txt
maco
malduck
smda
loguru
yara-python
